quantum_service/requirements.txt
flask==2.0.1
werkzeug==2.0.3
numpy>=1.25.0,<3.0
python-dotenv>=0.21.0
gunicorn==20.1.0
flask-cors==3.0.10
qiskit==1.1.1
qiskit-machine-learning==0.7.0
scikit-learn==1.3.0
pandas==2.0.3
urllib3~=2.0 